requirements.txt
appdirs==1.4.0
appnope==0.1.0
Babel==2.3.4
backports.shutil-get-terminal-size==1.0.0
cliff==2.1.0
cmd2==0.6.8
cwlref-runner==1.0
debtcollector==1.6.0
decorator==4.0.10
funcsigs==1.0.2
functools32==3.2.3.post2
ipython==5.0.0
ipython-genutils==0.1.0
iso8601==0.1.11
jsonpatch==1.14
jsonpointer==1.10
jsonschema==2.5.1
keystoneauth1==2.9.0
monotonic==1.1
msgpack-python==0.4.7
netaddr==0.7.18
netifaces==0.10.4
openstacksdk==0.9.0
os-client-config==1.18.0
oslo.config==3.13.0
oslo.i18n==3.8.0
oslo.serialization==2.11.0
oslo.utils==3.15.0
pathlib2==2.1.0
pbr==1.10.0
pexpect==4.2.0
pickleshare==0.7.2
pika==0.10.0
positional==1.1.1
prettytable==0.7.2
prompt-toolkit==1.0.3
ptyprocess==0.5.1
Pygments==2.1.3
pyparsing==2.1.5
python-cinderclient==1.8.0
python-glanceclient==2.1.0
python-keystoneclient==3.2.0
python-novaclient==4.1.0
python-openstackclient==2.6.0
pytz==2016.6
PyYAML==3.11
requests==2.10.0
requestsexceptions==1.1.3
rfc3986==0.3.1
simplegeneric==0.8.1
simplejson==3.8.2
six==1.10.0
stevedore==1.16.0
traitlets==4.2.2
unicodecsv==0.14.1
warlock==1.3.0
wcwidth==0.1.7
wrapt==1.10.8
